rename: card 폴더 components 하위로 이동 [feat/21]
