+++ src/components/card/contents/ContentText.js
